Main structure of repo, Makefile, README.md for project
Nothing implemented yet, although it is known what to do.

Some hesitations on macro names, I initially wanted to define them as
for example \texdimincm_[,u,d,safeu,safed] with _ of non letter catcode,
so in fact \texdimincm and then dispatch but changed my mind for use
case if _ is of catcode letter which would then need to be kept
separated.