Incredible... now 5% better only by turning on the wifi !
As well as having mail and firefox in the background.

This is very very puzzling, but benchmarking with \pdfelapsedtime as
**always** been very very very strange, and I did have puzzling things
in past, but not something like this:

1. first batch, new/old gains 5%

2. I repeat, new/old gains only 2% or 3%. Only difference: I turned off
my laptop wi-fi !!!!

3. I turn back on the wi-fi, and here we go, I again find a **more**
elevated new/old gain...

Notice here we are not talking about some overall speed change, but a
relative one between TeX macros !
Nothing changed in the code. But the benchmark outputs changed significantly...

	modified:   benchmark_withunit_random-2795649.txt
	modified:   benchmark_withunit_random-67740550.txt
	modified:   benchmark_withunit_random-70935444.txt
	modified:   benchmark_withunit_random-92545313.txt